proj_v1/requirements.txt
matplotlib==3.10.8
mysql_connector_repackaged==0.3.1
numpy==2.4.1
pandas==2.3.3
scikit_learn==1.8.0
seaborn==0.13.2
